--- src/components/Profile.jsx
@@ -1,18 +1,0 @@
-import { useAuth } from "./Auth"
-import { useNavigate } from "react-router-dom"
-
-export const Profile = () => {
-    const auth = useAuth()
-    const navigate = useNavigate()
-
-    const handleLogout = () => {
-        auth.logout()
-        navigate('/')
-    }
-    return (
-    <div>
-        Welcome {auth.user}
-        <button onClick={handleLogout}>Logout</button>
-    </div>
-    )
-}
