+++ src/components/profile/profile.css
@@ -1,0 +1,7 @@
+.profilepage__topbar{
+    padding: 2rem;
+}
+
+.profilepage__topbar button{
+    margin-left: 2rem;
+}
