+++ src/index.css
@@ -48,7 +48,12 @@ a:active{
 }
 
 .btn{
+  padding: 1rem;
   font-weight: bold;
+  border-radius: 0;
+  border: 1px solid var(--color-text);
+  background-color: var(--color-text);
+  color: var(--color-text-light);
   
   transition-duration: 150ms;
   transition-property: all;
@@ -56,9 +61,9 @@ a:active{
 }
 
 .btn-form{
-  padding: 1rem;
-  border-radius: 1px solid var(--color-text);
+  border: 1px solid var(--color-text);
   background-color: var(--color-bg);
+  color: var(--color-text);
 }
 
 .btn-form:hover,
